+++ fiddles/nativeScript/fiddle-0009-SqliteDb/fifteenpuzzle/app/app.js
@@ -1,0 +1,2 @@
+var application = require("application");
+application.start({ moduleName: "view/main-page" });